+++ tachyon/c/BUILD.bazel
@@ -52,7 +52,7 @@ TPLS = [
 ]
 
 CURVES = [
-    "bls/bls12_381",
+    "bls12/bls12_381",
     "bn/bn254",
 ]
