+++ tachyon/c/math/elliptic_curves/bls12/bls12_381/BUILD.bazel
@@ -19,6 +19,6 @@ generate_ec_points(
     name = "bls12_381",
     fq_limb_nums = 6,
     fr_limb_nums = 4,
-    g1_deps = ["//tachyon/math/elliptic_curves/bls/bls12_381:g1"],
-    g1_gpu_deps = ["//tachyon/math/elliptic_curves/bls/bls12_381:g1_gpu"],
+    g1_deps = ["//tachyon/math/elliptic_curves/bls12/bls12_381:g1"],
+    g1_gpu_deps = ["//tachyon/math/elliptic_curves/bls12/bls12_381:g1_gpu"],
 )
