+++ tachyon/c/math/elliptic_curves/generator/generator_util.cc
@@ -8,7 +8,7 @@ std::string GetLocation(std::string_view type) {
   if (type == "bn254") {
     return "bn/bn254";
   } else if (type == "bls12_381") {
-    return "bls/bls12_381";
+    return "bls12/bls12_381";
   }
   NOTREACHED() << "Unsupported type: " << type;
   return "";
