+++ tachyon/cc/BUILD.bazel
@@ -60,7 +60,7 @@ tachyon_cc_shared_library(
     soversion = VERSION,
     tags = ["manual"],
     deps = [
-        "//tachyon/cc/math/elliptic_curves/bls/bls12_381:g1",
+        "//tachyon/cc/math/elliptic_curves/bls12/bls12_381:g1",
         "//tachyon/cc/math/elliptic_curves/bn/bn254:g1",
     ],
 )
