+++ tachyon/cc/api.h
@@ -1,9 +1,9 @@
 #ifndef TACHYON_CC_API_H_
 #define TACHYON_CC_API_H_
 
-#include "tachyon/cc/math/elliptic_curves/bls/bls12_381/fq.h"
-#include "tachyon/cc/math/elliptic_curves/bls/bls12_381/fr.h"
-#include "tachyon/cc/math/elliptic_curves/bls/bls12_381/g1.h"
+#include "tachyon/cc/math/elliptic_curves/bls12/bls12_381/fq.h"
+#include "tachyon/cc/math/elliptic_curves/bls12/bls12_381/fr.h"
+#include "tachyon/cc/math/elliptic_curves/bls12/bls12_381/g1.h"
 #include "tachyon/cc/math/elliptic_curves/bn/bn254/fq.h"
 #include "tachyon/cc/math/elliptic_curves/bn/bn254/fr.h"
 #include "tachyon/cc/math/elliptic_curves/bn/bn254/g1.h"
