+++ tachyon/cc/math/BUILD.bazel
@@ -3,7 +3,7 @@ package(default_visibility = ["//visibility:public"])
 filegroup(
     name = "math_hdrs",
     srcs = [
-        "//tachyon/cc/math/elliptic_curves/bls/bls12_381:bls12_381_hdrs",
+        "//tachyon/cc/math/elliptic_curves/bls12/bls12_381:bls12_381_hdrs",
         "//tachyon/cc/math/elliptic_curves/bn/bn254:bn254_hdrs",
     ],
 )
