+++ tachyon/cc/math/elliptic_curves/bls12/bls12_381/BUILD.bazel
@@ -16,7 +16,7 @@ generate_ec_points(
     fq_limb_nums = 6,
     fr_limb_nums = 4,
     g1_deps = [
-        "//tachyon/c/math/elliptic_curves/bls/bls12_381:g1",
-        "//tachyon/math/elliptic_curves/bls/bls12_381:g1",
+        "//tachyon/c/math/elliptic_curves/bls12/bls12_381:g1",
+        "//tachyon/math/elliptic_curves/bls12/bls12_381:g1",
     ],
 )
