+++ tachyon/crypto/hashes/sponge/poseidon/BUILD.bazel
@@ -51,7 +51,7 @@ tachyon_cc_unittest(
     deps = [
         ":poseidon",
         ":poseidon_config",
-        "//tachyon/math/elliptic_curves/bls/bls12_381:fr",
+        "//tachyon/math/elliptic_curves/bls12/bls12_381:fr",
         "//tachyon/math/elliptic_curves/bn/bn254:fr",
     ],
 )
