+++ tachyon/crypto/hashes/sponge/poseidon/grain_lfsr_unittest.cc
@@ -7,7 +7,7 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/math/elliptic_curves/bls/bls12_381/fr.h"
+#include "tachyon/math/elliptic_curves/bls12/bls12_381/fr.h"
 
 namespace tachyon::crypto {
