+++ tachyon/crypto/hashes/sponge/poseidon/poseidon_config_unittest.cc
@@ -13,7 +13,7 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/math/elliptic_curves/bls/bls12_381/fr.h"
+#include "tachyon/math/elliptic_curves/bls12/bls12_381/fr.h"
 #include "tachyon/math/elliptic_curves/bn/bn254/fr.h"
 
 namespace tachyon::crypto {
