+++ tachyon/math/elliptic_curves/msm/algorithms/pippenger/BUILD.bazel
@@ -50,7 +50,7 @@ tachyon_cc_unittest(
     ],
     deps = [
         ":pippenger_adapter",
-        "//tachyon/math/elliptic_curves/bls/bls12_381:g1",
+        "//tachyon/math/elliptic_curves/bls12/bls12_381:g1",
         "//tachyon/math/elliptic_curves/bn/bn254:g1",
         "//tachyon/math/elliptic_curves/msm/test:msm_test_set",
     ],
