+++ tachyon/math/elliptic_curves/msm/glv_unittest.cc
@@ -2,8 +2,8 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/math/elliptic_curves/bls/bls12_381/g1.h"
-#include "tachyon/math/elliptic_curves/bls/bls12_381/g2.h"
+#include "tachyon/math/elliptic_curves/bls12/bls12_381/g1.h"
+#include "tachyon/math/elliptic_curves/bls12/bls12_381/g2.h"
 #include "tachyon/math/elliptic_curves/bn/bn254/g1.h"
 #include "tachyon/math/elliptic_curves/bn/bn254/g2.h"
 #include "tachyon/math/elliptic_curves/point_conversions.h"
