+++ tachyon/node/math/math.cc
@@ -1,10 +1,10 @@
 #include "tachyon/node/math/math.h"
 
-#include "tachyon/math/elliptic_curves/bls/bls12_381/g1.h"
+#include "tachyon/math/elliptic_curves/bls12/bls12_381/g1.h"
 #include "tachyon/math/elliptic_curves/bn/bn254/g1.h"
-#include "tachyon/node/math/elliptic_curves/bls/bls12_381/fq.h"
-#include "tachyon/node/math/elliptic_curves/bls/bls12_381/fr.h"
-#include "tachyon/node/math/elliptic_curves/bls/bls12_381/g1.h"
+#include "tachyon/node/math/elliptic_curves/bls12/bls12_381/fq.h"
+#include "tachyon/node/math/elliptic_curves/bls12/bls12_381/fr.h"
+#include "tachyon/node/math/elliptic_curves/bls12/bls12_381/g1.h"
 #include "tachyon/node/math/elliptic_curves/bn/bn254/fq.h"
 #include "tachyon/node/math/elliptic_curves/bn/bn254/fr.h"
 #include "tachyon/node/math/elliptic_curves/bn/bn254/g1.h"
